requirements.txt
streamlit
google-generativeai
pillow
pandas
numpy
opencv-python
opencv-python-headless
pytesseract
pdf2image
paddlepaddle
paddleocr
gTTS
streamlit-mic-recorder
streamlit-TTS
sqlalchemy
plotly
python-dotenv
